Utils fix
