+++ pwchem/utils/__init__.py
@@ -1,2 +1,1 @@
 from .utils import *
-from .utilsFasta import *
