+++ pwchem/utils/utils.py
@@ -393,4 +393,22 @@ def calculateDistance(coord1, coord2):
         dist += (c1-c2) ** 2
     return dist ** (1/2)
 
+def parseFasta(fastaFn, combined=True):
+  '''Parses a combined fasta file and returns a dictionary {seqId: sequence}'''
+  fastaDic = {}
+  seqId, seq = '', ''
+  with open(fastaFn) as f:
+    for line in f:
+      if line.startswith('>'):
+        if seqId != '':
+          fastaDic[seqId] = seq
+          if not combined:
+            # Return just the first sequence if asked for a single fasta
+            return fastaDic
+        seqId = line.strip()[1:]
+      elif line.strip() != '':
+        seq += line.strip()
+  fastaDic[seqId] = seq
+  return fastaDic
+
